clangd, style, fixes

+++ src/StateCommon.h
@@ -1,6 +1,7 @@
 #ifndef STATE_COMMON_H_
 #define STATE_COMMON_H_
+#include <cstdint>
 
-using CharState = int;
+using CharState = uint8_t;
 
 #endif
